cat-pictures-website/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.325
chromadb==0.4.18
openai==0.28.1
pydantic==2.5.2
python-multipart==0.0.6
